Renaming classes and files
